ai_service/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.responses import JSONResponse
from PIL import Image
import io
import numpy as np
import uvicorn

app = FastAPI(title="Xray-ECG AI Service", version="0.1.0")

try:
    # Optional CORS for browser clients
    from fastapi.middleware.cors import CORSMiddleware
    app.add_middleware(
        CORSMiddleware,
        allow_origins=["*"],
        allow_credentials=True,
        allow_methods=["*"],
        allow_headers=["*"],
    )
except Exception:
    # If CORS middleware not available, continue without it
    pass

# Placeholder class labels (14 common CXR findings)
CLASS_LABELS = [
    "Atelectasis","Cardiomegaly","Consolidation","Edema","Effusion",
    "Emphysema","Fibrosis","Hernia","Infiltration","Mass",
    "Nodule","Pleural_Thickening","Pneumonia","Pneumothorax"
]

@app.get("/health")
def health():
    return {"status": "ok"}

@app.post("/predict")
async def predict(cxr: UploadFile = File(...), ecg: UploadFile = File(...)):
    try:
        cxr_bytes = await cxr.read()
        # Basic validation the image is readable
        Image.open(io.BytesIO(cxr_bytes)).convert("RGB")
    except Exception:
        raise HTTPException(status_code=400, detail="Invalid CXR image")

    try:
        ecg_bytes = await ecg.read()
        # Accept .npy or csv-like bytes for demo. We'll just check non-empty.
        if len(ecg_bytes) == 0:
            raise ValueError("Empty ECG")
    except Exception:
        raise HTTPException(status_code=400, detail="Invalid ECG data")

    # Deterministic demo behavior based on CXR filename
    name = (cxr.filename or "").lower()

    if "image1" in name or "image2" in name:
        # Return a problem: set a couple of findings high
        probs_map = {label: 0.12 for label in CLASS_LABELS}
        probs_map["Cardiomegaly"] = 0.92
        probs_map["Effusion"] = 0.78
        findings = [
            {"finding": k, "confidence_score": float(v)}
            for k, v in probs_map.items() if v >= 0.5
        ]
        return JSONResponse({
            "success": True,
            "findings": findings,
            "probabilities": probs_map,
        })

    if "image3" in name:
        # Return no problem: keep all probabilities low
        probs_map = {label: 0.08 for label in CLASS_LABELS}
        return JSONResponse({
            "success": True,
            "findings": [],
            "probabilities": probs_map,
        })

    # Fallback demo prediction: deterministic pseudo-random scores based on sizes
    seed = len(cxr_bytes) + len(ecg_bytes)
    rng = np.random.default_rng(seed)
    probs = rng.random(len(CLASS_LABELS))
    findings = [
        {"finding": CLASS_LABELS[i], "confidence_score": float(p)}
        for i, p in enumerate(probs) if p >= 0.5
    ]
    return JSONResponse({
        "success": True,
        "findings": findings,
        "probabilities": {CLASS_LABELS[i]: float(p) for i, p in enumerate(probs)},
    })

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8001)
