ai_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pillow==11.0.0
numpy==2.1.2
python-multipart==0.0.9
